test-client/requirements.txt
redis>=5.0
redis-entraid>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
